// src/app/comment/types/comment-validation-error-code.enum.ts
export enum COMMENT_VALIDATION_ERROR_CODE {
  INVALID_SOURCE_REFERENCE = 'INVALID_SOURCE_REFERENCE',
  UNSUPPORTED_SOURCE_KIND = 'UNSUPPORTED_SOURCE_KIND',
  INVALID_LIMIT = 'INVALID_LIMIT',
  INVALID_DEPTH_RANGE = 'INVALID_DEPTH_RANGE',
  INVALID_PUBLISHED_RANGE = 'INVALID_PUBLISHED_RANGE',
  INVALID_SCORE_PERCENTILE = 'INVALID_SCORE_PERCENTILE',
  INVALID_SORT = 'INVALID_SORT',
}